Complete torch dataloading code
